requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.0
SQLAlchemy==2.0.36
alembic==1.13.2
httpx==0.27.2
Pillow==10.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
orjson==3.10.7
numpy
pillow
open-clip-torch
torch
torchvision
python-socketio>=5.11
